🐛 Fixes OPAP services bug
Instead of int sometimes they sent double..

+++ lib/models/statistics.dart
@@ -4,8 +4,7 @@
 
 import 'dart:convert';
 
-Statistics statisticsFromJson(String str) =>
-    Statistics.fromJson(json.decode(str));
+Statistics statisticsFromJson(String str) => Statistics.fromJson(json.decode(str));
 
 String statisticsToJson(Statistics data) => json.encode(data.toJson());
 
@@ -22,8 +21,7 @@ class Statistics {
 
   factory Statistics.fromJson(Map<String, dynamic> json) => Statistics(
         header: Header.fromJson(json["header"]),
-        numbers:
-            List<Number>.from(json["numbers"].map((x) => Number.fromJson(x))),
+        numbers: List<Number>.from(json["numbers"].map((x) => Number.fromJson(x))),
         bonusNumbers: List<Number>.from(
           json["bonusNumbers"].map((x) => Number.fromJson(x)),
         ),
@@ -73,7 +71,7 @@ class Header {
 
   factory Header.fromJson(Map<String, dynamic> json) => Header(
         dateFrom: json["dateFrom"],
-        dateTo: json["dateTo"],
+        dateTo: json["dateTo"].toInt(),
         drawCount: json["drawCount"],
       );
